+++ backend/app/api/src/crons.ts
@@ -4,7 +4,7 @@ import { Group, Organization, Payment, Registration, STPackage, Webshop } from '
 import { SQL } from '@stamhoofd/sql';
 import { PaymentMethod, PaymentProvider, PaymentStatus } from '@stamhoofd/structures';
 import { checkSettlements } from './helpers/CheckSettlements.js';
-import { OrganizationEmailService } from './services/OrganizationEmailService.js';
+import { OrganizationDNSService } from './services/OrganizationDNSService.js';
 import { PaymentService } from './services/PaymentService.js';
 import { RegistrationService } from './services/RegistrationService.js';
 import { STPackageService } from './services/STPackageService.js';
@@ -41,7 +41,7 @@ async function checkDNS() {
             console.log('[DNS] ' + organization.name);
         }
         try {
-            await OrganizationEmailService.updateDNSRecords(organization);
+            await OrganizationDNSService.updateDNSRecords(organization);
         } catch (e) {
             console.error(e);
         }
