+++ backend/app/api/src/endpoints/organization/dashboard/organization/SetOrganizationDomainEndpoint.ts
@@ -9,7 +9,7 @@ import NodeRSA from 'node-rsa';
 import { Formatter } from '@stamhoofd/utility';
 import { AuthenticatedStructures } from '../../../../helpers/AuthenticatedStructures.js';
 import { Context } from '../../../../helpers/Context.js';
-import { OrganizationEmailService } from '../../../../services/OrganizationEmailService.js';
+import { OrganizationDNSService } from '../../../../services/OrganizationDNSService.js';
 
 type Params = Record<string, never>;
 type Query = undefined;
@@ -170,7 +170,7 @@ export class SetOrganizationDomainEndpoint extends Endpoint<Params, Query, Body,
                 }));
             } else {
                 if (oldMailDomain) {
-                    organization.deleteAWSMailIdenitity(oldMailDomain).catch(console.error);
+                    OrganizationDNSService.deleteAWSMailIdenitity(organization, oldMailDomain).catch(console.error);
                 }
 
                 if (organization.serverMeta.privateDKIMKey && organization.serverMeta.publicDKIMKey) {
@@ -189,7 +189,7 @@ export class SetOrganizationDomainEndpoint extends Endpoint<Params, Query, Body,
         } else {
             // Validate DNS-records if not empty
             console.log('Validating domains');
-            await OrganizationEmailService.updateDNSRecords(organization);
+            await OrganizationDNSService.updateDNSRecords(organization);
         }
 
         console.log('Done.');
