+++ backend/shared/models/src/models/Organization.ts
@@ -1,5 +1,3 @@
-import type { GetEmailIdentityCommandOutput } from '@aws-sdk/client-sesv2';
-import { CreateEmailIdentityCommand, DeleteEmailIdentityCommand, GetEmailIdentityCommand, PutEmailIdentityFeedbackAttributesCommand, PutEmailIdentityMailFromAttributesCommand, SESv2Client } from '@aws-sdk/client-sesv2';
 import { column, Database } from '@simonbackx/simple-database';
 import { SimpleError } from '@simonbackx/simple-errors';
 import { I18n } from '@stamhoofd/backend-i18n/I18n';
@@ -352,167 +350,6 @@ export class Organization extends QueryableModel {
             createdAt: this.createdAt,
             hasFutureEvents: this.hasFutureEvents,
         });
-    }
-
-    static get forbiddenEmailDomains() {
-        return [
-            STAMHOOFD.domains.dashboard,
-            ...Object.values(STAMHOOFD.domains.defaultBroadcastEmail ?? {}),
-            ...Object.values(STAMHOOFD.domains.defaultTransactionalEmail ?? {}),
-        ];
-    }
-
-    async deleteAWSMailIdenitity(mailDomain: string) {
-        // Protect specific domain names
-        if (Organization.forbiddenEmailDomains.includes(mailDomain.toLowerCase())) {
-            return;
-        }
-
-        if (STAMHOOFD.environment !== 'production') {
-            // Temporary ignore this
-            return;
-        }
-
-        const client = new SESv2Client({});
-
-        // Check if mail identitiy already exists..
-        try {
-            const cmd = new GetEmailIdentityCommand({
-                EmailIdentity: mailDomain,
-            });
-            const result = await client.send(cmd);
-
-            if (result.VerifiedForSendingStatus === true) {
-                console.log('Cant delete AWS mail idenitiy @' + this.id + ' for ' + mailDomain + ': already validated and might be in use by other organizations');
-                return;
-            }
-
-            console.log('Deleting AWS mail identity @' + this.id + ' for ' + mailDomain);
-
-            const deleteCmd = new DeleteEmailIdentityCommand({
-                EmailIdentity: mailDomain,
-            });
-
-            await client.send(deleteCmd);
-            console.log('Deleted AWS mail idenitiy @' + this.id + ' for ' + this.privateMeta.mailDomain);
-        } catch (e) {
-            console.error('Could not delete AWS email identitiy @' + this.id + ' for ' + this.privateMeta.mailDomain);
-            console.error(e);
-        }
-    }
-
-    /**
-     * Create or update the AWS mail idenitiy and also update the active state of the mailDomain
-     */
-    async updateAWSMailIdenitity() {
-        if (this.privateMeta.mailDomain === null) {
-            return;
-        }
-
-        // Protect specific domain names
-        if (['stamhoofd.be', 'stamhoofd.nl', 'stamhoofd.shop', 'stamhoofd.app', 'stamhoofd.email'].includes(this.privateMeta.mailDomain)) {
-            console.error('Tried to validate AWS mail identity with protected domains @' + this.id);
-            this.privateMeta.mailDomainActive = false;
-            return;
-        }
-
-        if (STAMHOOFD.environment !== 'production') {
-            // Temporary ignore this
-            this.privateMeta.mailDomainActive = true;
-            return;
-        }
-
-        const client = new SESv2Client({});
-        const expectedConfigurationSetName = Formatter.slug(STAMHOOFD.platformName + '-domains');
-
-        // Check if mail identitiy already exists..
-        let exists = false;
-        let existing: GetEmailIdentityCommandOutput | undefined = undefined;
-        try {
-            const cmd = new GetEmailIdentityCommand({
-                EmailIdentity: this.privateMeta.mailDomain,
-            });
-
-            existing = await client.send(cmd);
-            exists = true;
-
-            console.log('AWS mail idenitiy exists already: just checking the verification status in AWS @' + this.id);
-
-            if (existing.ConfigurationSetName !== expectedConfigurationSetName) {
-                // Not allowed to use this identity
-                this.privateMeta.mailDomainActive = false;
-                console.error('Organization is not allowed to use email identity ' + this.privateMeta.mailDomain + ' @' + this.id + ', got ' + existing.ConfigurationSetName);
-                return;
-            }
-
-            this.privateMeta.mailDomainActive = existing.VerifiedForSendingStatus ?? false;
-
-            if (existing.VerifiedForSendingStatus !== true) {
-                console.error('Not validated @' + this.id);
-            }
-
-            if (existing.VerifiedForSendingStatus !== true && existing.DkimAttributes?.Status === 'FAILED') {
-                console.error('AWS failed to verify DKIM records. Triggering a forced recheck @' + this.id);
-
-                const deleteCmd = new DeleteEmailIdentityCommand({
-                    EmailIdentity: this.privateMeta.mailDomain,
-                });
-                await client.send(deleteCmd);
-
-                // Recreate it immediately
-                exists = false;
-            }
-        } catch (e) {
-            console.error(e);
-        }
-
-        if (!exists) {
-            console.log('Creating email identity in AWS SES...');
-
-            const cmd = new CreateEmailIdentityCommand({
-                EmailIdentity: this.privateMeta.mailDomain,
-                ConfigurationSetName: expectedConfigurationSetName,
-                DkimSigningAttributes: {
-                    DomainSigningPrivateKey: this.serverMeta.privateDKIMKey!,
-                    DomainSigningSelector: Formatter.slug(STAMHOOFD.platformName),
-                },
-                Tags: [
-                    {
-                        Key: 'OrganizationId',
-                        Value: this.id,
-                    },
-                    {
-                        Key: 'Environment',
-                        Value: STAMHOOFD.environment ?? 'Unknown',
-                    },
-                ],
-            });
-
-            const result = await client.send(cmd);
-            this.privateMeta.mailDomainActive = result.VerifiedForSendingStatus ?? false;
-
-            // Disable email forwarding of bounces and complaints
-            // We handle this now with the configuration set
-            const putFeedbackCmd = new PutEmailIdentityFeedbackAttributesCommand({
-                EmailIdentity: this.privateMeta.mailDomain,
-                EmailForwardingEnabled: false,
-            });
-
-            await client.send(putFeedbackCmd);
-        }
-
-        if (this.privateMeta.mailFromDomain && (!exists || (existing && (!existing.MailFromAttributes || existing.MailFromAttributes.MailFromDomain !== this.privateMeta.mailFromDomain)))) {
-            // Also set a from domain, to fix SPF
-            console.log('Setting mail from domain: ' + this.privateMeta.mailFromDomain + ' for ' + this.id);
-
-            const cmd = new PutEmailIdentityMailFromAttributesCommand({
-                EmailIdentity: this.privateMeta.mailDomain,
-                BehaviorOnMxFailure: 'USE_DEFAULT_VALUE',
-                MailFromDomain: this.privateMeta.mailFromDomain,
-            });
-
-            await client.send(cmd);
-        }
     }
 
     /**
